Add files for the reports page and moved emoji.ts to the app folder
