+++ client/src/app/home/home.service.spec.ts
@@ -1,7 +1,7 @@
 import {HttpClientTestingModule, HttpTestingController} from '@angular/common/http/testing';
 import {TestBed} from '@angular/core/testing';
 import {HttpClient} from '@angular/common/http';
-import {Emoji} from './emoji';
+import {Emoji} from '../emoji';
 import {HomeService} from './home.service'
 
 describe('Home service: ', () => {
